+++ src/core/Akka/Actor/SupervisorStrategy.cs
@@ -732,7 +732,7 @@ namespace Akka.Actor
     /// <summary>
     /// Collection of failures, used to keep track of how many times a given actor has failed.
     /// </summary>
-    [Obsolete("Use List of Akka.Actor.Status.Failure")]
+    [Obsolete("Use List of Akka.Actor.Status.Failure. Will be removed in v1.6")]
     public class Failures
     {
         /// <summary>
@@ -752,7 +752,7 @@ namespace Akka.Actor
     /// <summary>
     ///     Represents a single failure.
     /// </summary>
-    [Obsolete("Use Akka.Actor.Status.Failure")]
+    [Obsolete("Use Akka.Actor.Status.Failure. Will be removed in v1.6")]
     public class Failure
     {
         /// <summary>